src/api/requirements.txt
azure-functions
openai
azure-identity
python-dotenv
azure-cosmos>=4.5.0
azure-ai-projects

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
